updated k8s to typescript and a class
