+++ src/commands/gitops.js
@@ -1,16 +1,17 @@
-import { getResources, selectNamespace } from '../logic/k8s.js';
 import { preparePackage, processData } from '../logic/core.js';
+import { K8s } from '../logic/mod.ts';
 
 export async function gitopsCMD(namespace) {
+    const k8s = K8s();
     const dirPath = `./cf-support-gitops-${new Date().toISOString().replace(/[:.]/g, '-').replace(/\.\d{3}Z$/, 'Z')}`;
 
     if (!namespace) {
-        const selected = await selectNamespace();
+        const selected = await k8s.selectNamespace();
         namespace = selected;
     }
 
     console.log(`Gathering data in the '${namespace}' namespace for the GitOps Runtime`);
-    const k8sResources = getResources(namespace);
+    const k8sResources = k8s.getResources(namespace);
     await processData(dirPath, k8sResources);
     await preparePackage(dirPath);
 }
