+++ src/logic/mod.ts
@@ -1,3 +1,4 @@
 import { Codefresh } from './codefresh.ts';
+import { K8s } from './k8s.ts';
 
-export { Codefresh };
+export { Codefresh, K8s };
